Moved recording and slug helpers into shared package
